created --enable-serial option for configure which lets you set which PlatformSerial module to use created skeleton PlatformSerial module. By default the skeleton module is configured and built. created PlatformSerialFactory object
Every PlatformSerial module must define a PlatformSerialFactory object which will create an object using the IPlatformSerial interface, do any necessary initialization and return a shared pointer to this object. This separates the details of creating and initializing the hardware LED serial interface from
the rest of the code. This is the framework I will use to write a multi-platform LED string controller.