These changes implement the /bead osc method. For this to work the OSCServer now needs to know how many leds there are for each bead (8).
The GtkSim serial implementation also needs to know the leds per bead so it
can convert the buffer back to the visual representation of little circles on
the screen.  The physical serial implementations don't need this information.

The configure.ac changes check for JSONCPP. I will use this in a future update
to create a config file and move the hard coded values out of the object
initializers.

The Makefile.am change is left from before, and adds the options needed to
link GtkSim.